Support for adding query params to OIDC auth url in settings
In some circumstances[0], it is useful to provide extra parameters
to the authorization URL. When this is done, the resulting URL becomes
invalid as we end up with two '?' characters in the URL as the current
code expects there not to be any existing query params.

The requests library has a useful function that we can use to cleanly
provide extra params to a URL without having to worry about whether
or not there were any existing ones.

[0] https://www.keycloak.org/docs/latest/server_admin/#_client_suggested_idp
[1] https://www.markhneedham.com/blog/2019/01/11/python-add-query-parameters-url/

Change-Id: I65bade3cdae5bb714d7b0442a0e187dc8896aedf